+++ src/logLoss_.cpp
@@ -19,7 +19,7 @@ double mlogLoss_(NumericVector actual, NumericMatrix predicted) {
 
   NumericMatrix actualMat = NumericMatrix(Dimension(Rows, Cols));
 #pragma omp parallel for
-  for(int i = 0; i < Rows; ++i) {
+  for(int i = 0; i < (int) Rows; ++i) {
     actualMat(i, actual(i) - 1) = 1;
   }
